frontend/flyout: get rid of complex popovers (the flyout replaces them, at minimum), and move "server flyout" to its own file
